Corrected inflation chart

+++ financial-fortress-frontend/src/main/java/com/restapi/financialfortressfrontend/domain/dto/PortfolioValuesResponse.java
@@ -20,7 +20,7 @@ public class PortfolioValuesResponse {
     public Date date;
 
     @JsonProperty("entireValue")
-    public double entireValue;
+    public BigDecimal entireValue;
 
     @JsonProperty("goldValue")
     public BigDecimal goldValue;
